Refactor util functions into volCompareUtils
